move component and fix history
