+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "isunfa",
-  "version": "0.12.0+70",
+  "version": "0.12.0+69",
   "private": true,
   "type": "module",
   "engines": {
